quantum/requirements.txt
fastapi
uvicorn
numpy
seaborn
matplotlib
pandas
scikit-learn
dwave-ocean-sdk
fpdf
dwave-scikit-learn-plugin